Make Maigre.Style a Gtk.Style at the managed level
Properly matched the managed binding at the class and object level
to MaigreStyle (native) and Maigre.Style (managed). When the native
class_init is called, the class is also configured statically on
the managed side. When the native init is called, the object is
configured on the managed side.

The GType is now configured on the managed class, so it can be
a full binding. Essentially Maigre.Style complements the only partial
binding of the stock Gtk.Style, but is fully compatible now in
existing managed code.